+++ src/plugins/ncr/sagas/updateNodeFlow.js
@@ -1,13 +1,12 @@
-import { call, put, select } from 'redux-saga/effects';
-import { blur, getFormValues, reset, SubmissionError } from 'redux-form';
+import { call, put } from 'redux-saga/effects';
+import { reset, SubmissionError } from 'redux-form';
 import clearResponse from '@triniti/cms/plugins/pbjx/actions/clearResponse';
 import destroyEditor from '@triniti/cms/plugins/blocksmith/actions/destroyEditor';
-import formValues from '@triniti/cms/plugins/news/utils/formValues';
+import formData from '@triniti/cms/plugins/news/utils/formData';
 import NodeStatus from '@gdbots/schemas/gdbots/ncr/enums/NodeStatus';
 import startCase from 'lodash/startCase';
-import swal from 'sweetalert2';
-import isEqual from 'lodash/isEqual';
 
+import saveForceAfterSuccessFlow from './saveForceAfterSuccessFlow';
 import changeNodeFlow from './changeNodeFlow';
 
 export function* onAfterSuccessFlow({ config, history, match, resolve }) {
@@ -22,54 +21,7 @@ export function* onAfterSuccessFlow({ config, history, match, resolve }) {
   }
 
   if (config.shouldForceSave) {
-    const localStorageForm = yield formValues.get();
-    formValues.clear();
-
-    if (!Object.keys(localStorageForm).length) {
-      return;
-    }
-
-    const submittedValues = yield select(getFormValues(config.formName));
-    const serializedValues = yield JSON.parse(JSON.stringify(submittedValues));
-    const localValues = localStorageForm.values;
-
-    // for some reason a value could not be save/persisted then it's invalid
-    const unsavedFields = yield Object.keys(localValues)
-      .filter((fieldName) => !isEqual(localValues[fieldName], serializedValues[fieldName])
-        // exclude if comparing empty string to null
-        && !(localValues[fieldName] === '' && serializedValues[fieldName] === null));
-
-    const formErrors = yield Object.assign(
-      localStorageForm.asyncErrors || {},
-      localStorageForm.submitErrors,
-      localStorageForm.syncErrors,
-    );
-    const invalidFields = yield [].concat(unsavedFields, Object.keys(formErrors))
-      .filter((fieldName, index, self) => self.indexOf(fieldName) === index);
-
-    if (!invalidFields.length) {
-      return;
-    }
-
-    const values = localStorageForm.values;
-    const result = yield swal.fire({
-      title: 'All updates have been saved, except:',
-      html: `<strong>${invalidFields.sort().join(', ')}</strong>`,
-      showCancelButton: true,
-      confirmButtonText: 'Resolve',
-      cancelButtonText: 'Dismiss',
-      confirmButtonClass: 'btn btn-primary',
-      cancelButtonClass: 'btn btn-secondary',
-    });
-
-    if (result.value) {
-      let i = 0;
-      for (; i < invalidFields.length; i += 1) {
-        const fieldName = invalidFields[i];
-        const value = values[fieldName];
-        yield put(blur(config.formName, fieldName, value));
-      }
-    }
+    yield saveForceAfterSuccessFlow(config);
   }
 }
 
@@ -77,7 +29,7 @@ export function* onAfterFailureFlow({ config, reject }, error) {
   const message = typeof error.getMessage === 'function' ? error.getMessage() : error.message;
   yield call(reject, new SubmissionError({ _error: message }));
   if (config.shouldForceSave) {
-    formValues.clear();
+    formData.clear();
   }
 }
